refactoring - part 2
