+++ components/robotsports/local_ball_tracking/BUILD
@@ -16,6 +16,9 @@ cc_library(
     name = "implementation",
     hdrs = [
         "RobotsportsLocalBallTracking.hpp", # generated by MRA-codegen.py
+        "RobotsportsLocalBallTracking_preprocessing.hpp",
+        "RobotsportsLocalBallTracking_native_filter.hpp",
+        "RobotsportsLocalBallTracking_sequence_clustering.hpp",
         "constants_ball_model.hpp",
         "seq_clustering_balldef.hpp",
         "seq_clustering_ball_model.hpp",
@@ -23,6 +26,9 @@ cc_library(
     ],
     srcs = [
         "tick.cpp", # to be implemented by user, generated if not existing
+        "RobotsportsLocalBallTracking_preprocessing.cpp",
+        "RobotsportsLocalBallTracking_native_filter.cpp",
+        "RobotsportsLocalBallTracking_sequence_clustering.cpp","
         "seq_clustering_ball_model.cpp",
         "seq_clustering_best_uid.cpp",
     ],
