+++ components/robotsports/local_ball_tracking/CMakeLists.txt
@@ -3,6 +3,9 @@
 
 add_library(MRA-components-robotsports-local-ball-tracking
     tick.cpp
+    RobotsportsLocalBallTracking_preprocessing.cpp
+    RobotsportsLocalBallTracking_native_filter.cpp
+    RobotsportsLocalBallTracking_sequence_clustering.cpp
 	seq_clustering_best_uid.cpp
 	seq_clustering_ball_model.cpp
 )
