+++ components/robotsports/local_ball_tracking/RobotsportsLocalBallTracking_native_filter.hpp
@@ -1,0 +1,15 @@
+#ifndef MRA_ROBOTSPORTS_LOCAL_BALL_TRACKING_NATIVE_FILTER_HPP
+#define MRA_ROBOTSPORTS_LOCAL_BALL_TRACKING_NATIVE_FILTER_HPP
+
+// generated protobuf types from interface of this component
+#include "RobotsportsLocalBallTracking.hpp"
+#include "seq_clustering_balldef.hpp"
+
+void local_ball_tracking_native_filter(int nrBallsThisTime,
+                                       std::vector<ball_feature_t>& ballData,
+                                        const MRA::RobotsportsLocalBallTracking::ParamsType &params,
+                                        const MRA::RobotsportsLocalBallTracking::InputType &input,
+                                        MRA::RobotsportsLocalBallTracking::StateType &state,
+                                        MRA::RobotsportsLocalBallTracking::OutputType &output);
+
+#endif
